idk wip

+++ app/src/main/java/me/exeos/jjvm/vm/bytecode/insn/AbstractInstruction.java
@@ -2,9 +2,9 @@ package me.exeos.jjvm.vm.bytecode.insn;
 
 public abstract class AbstractInstruction {
 
-    public final int opcode;
+    public final byte opcode;
 
-    public AbstractInstruction(int opcode) {
+    public AbstractInstruction(byte opcode) {
         this.opcode = opcode;
     }
